Major rewrite and restructuring in prep for PR.
Essentially, I'm a lot wiser now (I think...)
So with Wen's overhaul of the concurrency primitives, we have Channels
but they block on get/put. This is useful in some scenarios and not so
useful in others. Hence, I made a more solid effort into making this
thing work/nicer to work with.

BufferedChannels have a single Queue, Pipes have 2.
Messages are gone since they made everything an unsafe nightmare, and
I'm not convinced they were ever a good idea (especially not after
working on parallel skeletons and using Types in general). I think now
I'd say "If you want to send multiple things, you can: just wrap them in
a custom type! :)"

Need to rewrite the tests to match, but then I believe it should be good
for opening a PR.